Add option to generate "spiral" fill for circles
This allows pen plotters (and other CNC machines) with narrower width
to generate path-filled circles (rather than just hollow circles) as
output. The GUI has a new button to switch this on/off as well.

The spiral path does give a much larger file size-- it is a simple
point-based path rather than an efficient spline -- however, it is
quite effective.